% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIbatch-deprecated.R
\name{NVIbatch-deprecated}
\alias{NVIbatch-deprecated}
\title{Deprecated Functions in Package NVIbatch}
\arguments{
\item{\dots}{(arguments)}
}
\value{
(results)
}
\description{
These functions are provided for compatibility with older
    versions of NVIbatch only, and may be defunct as soon as the next release.
    When possible, alternative functions are mentioned. Help pages for
    deprecated functions are available at \code{help("<function>-deprecated")}.
}
\details{
The argument \code{additional_info} in \code{save_log} was
    deprecated from v0.6.0 released 2024-##-##. The argument was renamed to
    the more meaningful \code{include_text}. If using the old argument,
    the input will be transferred to the new argument.
}
\examples{
\dontrun{
save_log(...) ### -- use argument \code{include_text} instead of \code{additional_info}.
}
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
\keyword{internal}
