% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NVIbatch-package.R
\docType{package}
\name{NVIbatch-package}
\alias{NVIbatch}
\alias{NVIbatch-package}
\title{NVIbatch: Tools to facilitate the running of R-scripts in batch mode at NVI}
\description{
Provides tools to facilitate the writing of batch R-scripts that should be run automatically at specific times.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/NorwegianVeterinaryInstitute/NVIbatch}
  \item Report bugs at \url{https://github.com/NorwegianVeterinaryInstitute/NVIbatch/issues}
}

}
\author{
\strong{Maintainer}: Petter Hopp \email{Petter.Hopp@vetinst.no} (\href{https://orcid.org/0000-0002-8695-0378}{ORCID})

Other contributors:
\itemize{
  \item Norwegian Veterinary Institute [copyright holder]
}

}
\keyword{internal}
