% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/old_NVIverse.R
\name{old_NVIverse}
\alias{old_NVIverse}
\title{Identify old packages within NVIverse}
\usage{
old_NVIverse(lib.loc = NULL, pkg = NVIpkgs, auth_token = NULL)
}
\arguments{
\item{lib.loc}{character vector describing the location of R library trees to search through-out, or NULL for all known trees (see .libPaths).}

\item{pkg}{NVIverse packages that should be checked. Only packages listed and installed are checked. Defaults to all NVIverse packages.}
}
\description{
old_NVIverse indicates packages within NVIverse that have a later version on the Github repository.
}
\details{
Only NVIverse packages that already are installed will be checked for even if more pkg are given as ibput to pkg.
}
