% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_log.R
\name{save_log}
\alias{save_log}
\title{Save and report log for scripts}
\usage{
save_log(
  log_file,
  log_path,
  save = TRUE,
  archive = NULL,
  email = TRUE,
  from = NULL,
  to = NULL,
  additional_info = NULL,
  smtp_server = NULL,
  ...
)
}
\arguments{
\item{log_file}{[\code{character(1)}]\cr
File name of the log file.}

\item{log_path}{[\code{character(1)}]\cr
Path for the log file.}

\item{save}{[\code{logical(1)}]\cr
Whether the log file should be saved or not. Defaults to \code{TRUE}.}

\item{archive}{[\code{character(1)}]\cr
The path to the directory to save the log file. Defaults to \code{NULL}.}

\item{email}{[\code{logical(1)}]\cr
Whether an email with a report of the status in the log and the log file
attached and the results file should be sent or not. Defaults to \code{TRUE}.}

\item{from}{[\code{character(1)}]\cr
The email address of the sender. Defaults to \code{NULL}.}

\item{to}{[\code{character}]\cr
The email address' of the recipients. Defaults to \code{NULL}.}

\item{additional_info}{[\code{character(1)}]\cr
Additional text to be written in the first part of the body of the email.
Defaults to \code{NULL}.}

\item{smtp_server}{[\code{character(1)}]\cr
The email server that sends the emails. Defaults to \code{NULL}.}

\item{\dots}{Other arguments to be passed to \code{\link{gather_messages}} and
\ifelse{html}{\code{\link[sendmailR:sendmail]{sendmailR::sendmail}}}{\code{sendmailR::sendmail}}.}
}
\value{
None. Saves the log after a script has been run
    and reports the error and warnings written in the log.
}
\description{
Reads the log file, identifies error messages and
    sends an email with the status for running the main script.
    The log file can be saved with the current date last in
    file name.
}
\details{
The input to the argument \code{params = } should be a list with
    parameters that is used by the rmarkdown document. The parameters must
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
