% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_NVIverse.R
\name{use_NVIverse}
\alias{use_NVIverse}
\title{Attach and if necessary install packages within \code{NVIverse}}
\usage{
use_NVIverse(
  pkg,
  auth_token = NULL,
  dependencies = NA,
  upgrade = FALSE,
  build = TRUE,
  build_manual = FALSE,
  build_vignettes = TRUE,
  ...
)
}
\arguments{
\item{pkg}{[\code{character}]\cr
Name of one or more \code{NVIverse} packages.}

\item{auth_token}{[\code{character}]\cr
To install \code{NVIconfig} a personal access token is needed. Generate a
personal access token (PAT) in "https://github.com/settings/tokens" and
supply to this argument. Defaults to \code{NULL}.}

\item{dependencies}{[\code{logical(1) | character}]\cr
The dependencies to check and eventually install. Can be
    a character vector (selecting from "Depends", "Imports", "LinkingTo",
    "Suggests", or "Enhances"), or a logical vector. \code{TRUE} is shorthand
    for c("Depends", "Imports", "LinkingTo", "Suggests"). \code{FALSE} is
    shorthand for no dependencies, i.e. just check this package, not its
    dependencies. \code{NA} is shorthand for c("Depends", "Imports", "LinkingTo")
    and is the default.}

\item{upgrade}{[\code{logical(1) | character(1)}]\cr
Should package dependencies be upgraded? One of c("ask", "always", "never").
    \code{TRUE} and \code{FALSE} are also accepted and correspond to
    "always" and "never" respectively. Defaults to \code{FALSE}.}

\item{build}{[\code{logical(1)}]\cr
If \code{TRUE} build the package before installing. Defaults to \code{TRUE}.}

\item{build_manual}{[\code{logical(1)}]\cr
If \code{FALSE}, don't build PDF manual ('--no-manual').
    Defaults to \code{FALSE}.}

\item{build_vignettes}{[\code{logical(1)}]\cr
If \code{FALSE}, don't build package vignettes ("--no-build-vignettes").
    Defaults to \code{TRUE}, but is set to \code{FALSE} if necessary packages
    for building vignettes are not installed.}

\item{\dots}{Other arguments to be passed to
\ifelse{html}{\code{\link[remotes:install_github]{remotes::install_github}}}{\code{remotes::install_github}}.}
}
\description{
First, \code{use_NVIverse} checks if the package is installed. If not
    already installed, the package will be installed. Thereafter, the package
    is attached using library.
}
\details{
Only packages within the \code{NVIverse} can be installed.
}
\examples{
use_NVIverse("NVIcheckmate")
use_NVIverse(pkg = c("NVIcheckmate", "NVIdb"))
}
