% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_pkg.R
\name{use_pkg}
\alias{use_pkg}
\title{Attach and if necessary install a package}
\usage{
use_pkg(pkg, repos = "https://cran.uib.no/", ...)
}
\arguments{
\item{pkg}{A vector with the name of one or more packages.}

\item{repos}{character vector, the base URL(s) of the repositories to use,
e.g., the URL of a Cran mirror such as "https://cloud.r-project.org". Can
be \code{NULL} to install from local files, directories or URLs: this will be
inferred by extension from pkgs if of length one.}

\item{\dots}{Other arguments to be passed to \code{install.packages}.}
}
\description{
First, \code{use_package} checks if the package is installed. If not
already installed, the package will be installed. Thereafter, the package
is attached using library.
}
\details{
Only packages available at Cran can be installed.
}
\examples{
use_pkg("checkmate")
use_pkg(pkg = c("checkmate", "devtools"))
}
