% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_rendered.R
\name{output_rendered}
\alias{output_rendered}
\title{Render an rmarkdown file and output the result file.}
\usage{
output_rendered(
  input,
  output_file,
  output_dir,
  intermediates_dir = tempdir(),
  params,
  display = FALSE,
  email = FALSE,
  from = NULL,
  to = NULL,
  subject = NULL,
  email_text = NULL,
  ...
)
}
\arguments{
\item{input}{The path to the rmarkdown document.}

\item{output_file}{[\code{character(1)}]. The name of the output file.}

\item{output_dir}{[\code{character(1)}]. The directory to save the output file.}

\item{intermediates_dir}{[\code{character(1)}]. The directory to save
intermediate files made by rmarkdown::render. Defaults to tempdir().}

\item{params}{[\code{list}]. List of parameters to be passed to the rmarkdown
document. The parameters must have been defined in the YAML-section of the
rmarkdown document.}

\item{display}{[\code{logical(1)} |\code{character(1)}]. If \code{FALSE}, don't
display the results file. Can also be "browser" for the default browser
or "viewer" for the R studio viewer. Defaults to \code{FALSE}.}

\item{email}{[\code{logical(1)}]. Whether an email with the results file
should be sent or not. Defaults to \code{FALSE}.}

\item{from}{[\code{character(1)}]. The email address of the sender.}

\item{to}{[\code{character}]. The email address' of the recipients.}

\item{subject}{[\code{character(1)}]. Text in the subject line of the email.
Defaults to the filename.}

\item{email_text}{[\code{character(1)}]\. Text to be written in the body of
the email.}

\item{\dots}{Other arguments to be passed to \code{rmarkdown::render} and
\code{sendmailR::sendmail}.}
}
\description{
Render an rmarkdown file, save the result file and eventually
display the output in the browser or send the output to one or more
email recipients.
}
\details{
The output can only be displayed in a browser or the R studio viewer
in an interaktive session. If you chose to display the output in the
R studio viewer, but it is not an R studio session, the output will
instead be output in the browser.

\if{html}{\out{<div class="sourceCode">}}\preformatted{An email with the results file can be sent to one or more recipients.
}\if{html}{\out{</div>}}
}
