% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_log.R
\name{save_log}
\alias{save_log}
\title{Save and report log for scripts}
\usage{
save_log(
  log_file,
  log_path,
  save = TRUE,
  archive = NULL,
  email = TRUE,
  from = NULL,
  to = NULL,
  include_text = NULL,
  attach_object = NULL,
  smtp_server = NULL,
  ...
)
}
\arguments{
\item{log_file}{[\code{character(1)}]\cr
File name of the log file.}

\item{log_path}{[\code{character(1)}]\cr
Path for the log file.}

\item{save}{[\code{logical(1)}]\cr
Whether the log file should be saved or not. Defaults to \code{TRUE}.}

\item{archive}{[\code{character(1)}]\cr
The path to the directory to save the log file. Defaults to \code{NULL}.}

\item{email}{[\code{logical(1)}]\cr
Whether an email with a report of the status in the log and the log file
attached and the results file should be sent or not. Defaults to \code{TRUE}.}

\item{from}{[\code{character(1)}]\cr
The email address of the sender. Defaults to \code{NULL}.}

\item{to}{[\code{character}]\cr
The email address' of the recipients. Defaults to \code{NULL}.}

\item{include_text}{[\code{character(1)}]\cr
Text to include in the first part of the body of the email. Defaults to
\code{NULL}.}

\item{attach_object}{[\code{character}]\cr
Full path and file name of object(s) to attach to the email. Defaults to
\code{NULL}.}

\item{smtp_server}{[\code{character(1)}]\cr
The email server that sends the emails. Defaults to \code{NULL}.}

\item{\dots}{Other arguments to be passed to \code{\link{gather_messages}} and
\ifelse{html}{\code{\link[sendmailR:sendmail]{sendmailR::sendmail}}}{\code{sendmailR::sendmail}}.}
}
\value{
None. Saves the log after a script has been run
    and reports the error and warnings written in the log.
}
\description{
Reads the log file, identifies error messages and
    sends an email with the status for running the main script.
    The log file can be saved with the current date last in
    file name.
}
\details{
The function collects warnings and error messages from
    log file when a script has been run. This is dependant on that
    the script has been set up to produce the log file. There
    the name and path of the log file will have been given.

The log file will be renamed so that the current date in the
    format "yyyymmdd" is included last in the file name before the
    file extension. This file should be permanently saved in an
    archive directory for example a subdirectory named "batchlog".

If an email should report the status, an email is produced. The
    subject of the email indicates whether the script run without
    problems or if warnings or error messages were produced. The
    log file is attached the email. If there are warnings or error
    messages, these are written in the email text.

It is possible include an short message first in the email text
    by giving it as input to the argument \code{include_text}. Such
    a message can give more information of the status of running the
    script. Such a message can be produced by the script and saved.
    Thereafter, the message can be fetched and be used as input to
    \code{include_text}.
}
\author{
Petter Hopp Petter.Hopp@vetinst.no
}
