% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_package.R
\name{use_package}
\alias{use_package}
\title{Install and attach a package available at Cran}
\usage{
use_package(pkg, repos = "https://cran.uib.no/", ...)
}
\arguments{
\item{pkg}{The name of the package.}

\item{repos}{character vector, the base URL(s) of the repositories to use,
e.g., the URL of a CRAN mirror such as "https://cloud.r-project.org". Can
be NULL to install from local files, directories or URLs: this will be
inferred by extension from pkgs if of length one.}

\item{\dots}{Other arguments to be passed to install.packages.}
}
\description{
First, use_package checks if the package is installed. If not
already installed, the package will be installed. Thereafter, the package
is attached using library.
}
\details{
Only packages available at Cran can be installed.
}
\examples{
use_package("checkmate")
}
